Install the FakeClockModule in MiskTestingServiceModule
